requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625037499685/work
certifi==2021.5.30
Django @ file:///tmp/build/80754af9/django_1622711766781/work
django-simpleui==2021.6.2
djangorestframework==3.12.4
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
wincertstore==0.2

requests~=2.25.1
django~=3.2.4